chore: seperated user location function into services, rehomed components
