+++ src/components/Tour/components/MarkerMap/components/MainMap/components/UserLocation.tsx
@@ -1,20 +1,8 @@
 import { useAtomValue } from "jotai";
-import { divIcon } from "leaflet";
 import { Marker, Pane } from "react-leaflet";
 
-import { getUserLocationAtom } from "./../atoms";
-
-function headingMarkerFactory() {
-  const img = document.createElement("img");
-  img.setAttribute("src", "/icons/user_location.svg");
-
-  return divIcon({
-    html: img,
-    className: "",
-    iconSize: [40, 40],
-    iconAnchor: [20, 20],
-  });
-}
+import { getUserLocationAtom } from "../../../../../../../atoms";
+import { headingMarkerFactory } from "../../../../../../../services";
 
 function UserLocation() {
   const userLocation = useAtomValue(getUserLocationAtom);
